examples/multi-agent/requirements.txt
# Multi-Agent System with CrewAI
# Python 3.9+

# CrewAI Framework
crewai>=0.1.0
crewai-tools>=0.1.0

# LangChain Integration
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# LLM Providers
anthropic>=0.7.0
openai>=1.3.0  # Optional: for embeddings

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.0

# Optional: Web Search
google-search-results>=2.4.2  # For Serper
duckduckgo-search>=3.9.0      # Alternative search

# Optional: Document Processing
pypdf>=4.0.0
python-docx>=0.8.0
markdown>=3.5.0

# Development
langsmith>=0.0.68
